requirements.txt
click
click-repl
ipython
mypy
mypy_extensions
nltk
pyspellchecker
pytest
pytest-mypy
scrapy
